+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - fix download json for text and analyse
+- fix download json for whois
 - use collapse for analyse and text
 
 ### Removed